requirements.txt
django
Pillow
stripe
